# Optimize `lagrange_poly()`