--- a/docs/index.rst
+++ b/docs/index.rst
@@ -140,6 +140,7 @@ If this library was useful to you in your research, please cite us. Following th
    :hidden:
 
    release-notes/versioning.rst
+   release-notes/v0.1.2.md
    release-notes/v0.1.1.md
    release-notes/v0.1.0.md
    release-notes/v0.0.33.md