--- a/tests/test_galois_lfsr.py
+++ b/tests/test_galois_lfsr.py
@@ -45,7 +45,7 @@ def test_repr():
 def test_str():
     c = galois.primitive_poly(7, 4)
     lfsr = galois.GLFSR(c.reverse())
-    assert str(lfsr) == "Galois LFSR:\n  field: GF(7)\n  feedback_poly: 5x^4 + 3x^3 + x^2 + 1\n  characteristic_poly: x^4 + x^2 + 3x + 5\n  taps: [2, 4, 6, 0]\n  order: 4\n  state: [1, 1, 1, 1]\n  initial_state: [1, 1, 1, 1]"
+    assert str(lfsr) == "Galois LFSR:\n  field: GF(7)\n  feedback_poly: 5x^4 + 3x^3 + x^2 + 1\n  characteristic_poly: x^4 + x^2 + 3x + 5\n  taps: [2 4 6 0]\n  order: 4\n  state: [1 1 1 1]\n  initial_state: [1 1 1 1]"
 
 
 @pytest.mark.parametrize("c", CHARACTERISTIC_POLYS)