@Popeyef5 2023-01-03
Awesome, thank you for replying so quickly and sorry for the late reply. I got caight up with work and had to put the project on hold for a bit. 

Regarding this, my issue is that I'm working with large prime finite fields, so JIT compilation isn't an option. I'm looking into different alternatives because it's taking prohibitely long to interpolate.